API

API

Types

ValueShapes.AbstractScalarShape

ValueShapes.AbstractValueShape

ValueShapes.ArrayShape

ValueShapes.ConstValueShape

ValueShapes.NamedTupleShape

ValueShapes.ScalarShape

ValueShapes.ValueAccessor

Functions

ValueShapes.default_datatype

ValueShapes.totalndof

ValueShapes.valshape

Documentation

ValueShapes.AbstractScalarShape — Type.

AbstractScalarShape{T} <: AbstractValueShape

source

ValueShapes.AbstractValueShape — Type.

abstract type AbstractValueShape

An AbstractValueShape combines type and size information.

Subtypes are defined for shapes of scalars (see ScalarShape), arrays (see ArrayShape), constant values (see ConstValueShape) and NamedTuples (see NamedTupleShape).

Subtypes of AbstractValueShape must support eltype, size and totalndof.

Value shapes can be used as constructors to generate values of the given shape with undefined content. If the element type of the shape is an abstract or union type, a suitable concrete type will be chosen automatically, if possible (see ValueShapes.default_datatype):

shape = ArrayShape{Real}(2,3)
A = shape(undef)
typeof(A) == Array{Float64,2}
size(A) == (2, 3)
valshape(A) == ArrayShape{Float64}(2,3)

Use

(shape::AbstractValueShape)(data::AbstractVector{<:Real})::eltype(shape)

to view a flat vector of anonymous real values as a value of the given shape:

data = [1, 2, 3, 4, 5, 6]
shape(data) == [1 3 5; 2 4 6]

In return,

Base.Vector{T}(undef, shape::AbstractValueShape)
Base.Vector(undef, shape::AbstractValueShape)

will create a suitable uninitialized vector of the right length to hold such flat data for the given shape. If no type T is given, a suitable data type will be chosen automatically.

When dealing with multiple vectors of flattened data, use

shape.(data::ArrayOfArrays.AbstractVectorOfSimilarVectors)

ValueShapes supports this via specialized broadcasting.

In return,

ArraysOfArrays.VectorOfSimilarVectors{T}(shape::AbstractValueShape)
ArraysOfArrays.VectorOfSimilarVectors(shape::AbstractValueShape)

will create a suitable vector (of length zero) of vectors that can hold flattened data for the given shape. The result will be a VectorOfSimilarVectors wrapped around a 2-dimensional ElasticArray. This way, all data is stored in a single contiguous chunk of memory.

source

ValueShapes.ArrayShape — Type.

ArrayShape{T,N} <: AbstractValueShape

Describes the shape of N-dimensional arrays of type T and a given size.

Constructor:

ArrayShape{T}(dims::NTuple{N,Integer}) where {T,N}
ArrayShape{T}(dims::Integer...) where {T}

e.g.

shape = ArrayShape{Real}(2, 3)

In addition to using the shape as a value constructor

size(shape(undef)) == (2, 3)
eltype(shape(undef)) == Float64

(see AbstractValueShape), a shape can also be used as an argument of certains array type constructors to explicitly construct standard Arrays or ElasticArrays:

using ElasticArrays

size(Array(undef, shape)) == (2, 3)
eltype(Array(undef, shape)) == Float64

size(ElasticArray(undef, shape)) == (2, 3)

See also the documentation of AbstractValueShape.

source

ValueShapes.ConstValueShape — Type.

ConstValueShape{T} <: AbstractValueShape

A ConstValueShape describes the shape of constant values of type T.

Constructor:

ConstValueShape(value)

value may be of arbitrary type, e.g. a constant scalar value or array:

ConstValueShape(4.2)
ConstValueShape([11 21; 12 22])

Shapes of constant values have zero degrees of freedom (see totalndof).

See also the documentation of AbstractValueShape.

source

ValueShapes.NamedTupleShape — Type.

NamedTupleShape{names,...} <: AbstractValueShape

Defines the shape of a NamedTuple (resp. set of variables, parameters, etc.).

Constructors:

NamedTupleShape(name1 = shape1::AbstractValueShape, ...)
NamedTupleShape(named_shapes::NamedTuple)

e.g.

shape = NamedTupleShape(
    a = ArrayShape{Real}(2, 3),
    b = ScalarShape{Real}(),
    c = ArrayShape{Real}(4)
)

Example:

shape = NamedTupleShape(
    a = ScalarShape{Real}(),
    b = ArrayShape{Real}(2, 3),
    c = ConstValueShape(42)
)
data = VectorOfSimilarVectors{Float64}(shape)
resize!(data, 10)
rand!(flatview(data))
table = shape.(data)
fill!(table.a, 4.2)
all(x -> x == 4.2, view(flatview(data), 1, :))

See also the documentation of AbstractValueShape.

source

ValueShapes.ScalarShape — Type.

ScalarShape{T} <: AbstractScalarShape{T}

An ScalarShape describes the shape of scalar values of a given type.

Constructor:

ScalarShape{T::Type}()

T may be an abstract type of Union, or a specific type, e.g.

ScalarShape{Real}()
ScalarShape{Integer}()
ScalarShape{Float32}()
ScalarShape{Complex}()

Scalar shapes may have a total number of degrees of freedom (see totalndof) greater than one, e.g. shapes of complex-valued scalars:

totalndof(ScalarShape{Real}()) == 1
totalndof(ScalarShape{Complex}()) == 2

See also the documentation of AbstractValueShape.

source

ValueShapes.ValueAccessor — Type.

ValueAccessor{S<:AbstractValueShape}

A value accessor provides a means to access a value with a given shape stored in a flat real-valued data vector with a given offset position.

Constructor:

ValueAccessor{S}(shape::S, offset::Int)

The offset is relative to the first index of a flat data array, so if the value is stored at the beginning of the array, the offset will be zero.

An ValueAccessor can be used to index into a given flat data array.

Example:

acc = ValueAccessor(ArrayShape{Real}(2,3), 2)
valshape(acc) == ArrayShape{Real,2}((2, 3))
data = [1, 2, 3, 4, 5, 6, 7, 8, 9]
data[acc] == [3 5 7; 4 6 8]

source

ValueShapes.totalndof — Function.

totalndof(shape::AbstractValueShape)

Get the total number of degrees of freedom of values of the given shape.

Equivalent to the length of a vector that would result from flattening the data into a sequence of real numbers, excluding any constant values.

source

ValueShapes.valshape — Function.

valshape(x)::AbstractValueShape
valshape(acc::ValueAccessor)::AbstractValueShape
valshape(d::Distributions.Distribution)::AbstractValueShape

Get the value shape of an arbitrary value, resp. the shape a ValueAccessor is based on, or the shape of the variates for a Distribution.

source

ValueShapes.default_datatype — Function.

ValueShapes.default_datatype(T::Type)

Return a default specific type U that is more specific than T, with U <: T.

e.g.

ValueShapes.default_datatype(Real) == Float64
ValueShapes.default_datatype(Complex) == Complex{Float64}

source