API

API

Types

ShapesOfVariables.AbstractValueShape

ShapesOfVariables.ArrayShape

ShapesOfVariables.ConstValueShape

ShapesOfVariables.ScalarShape

ShapesOfVariables.VarShapes

Functions

ShapesOfVariables.default_datatype

ShapesOfVariables.totalndof

Documentation

ShapesOfVariables.AbstractValueShape — Type.

abstract type AbstractValueShape

An AbstractValueShape combines type and size information, the combination of which is termed shape, here. Subtypes are defined for shapes of scalars (see ScalarShape), arrays (see ArrayShape) and constant values (see ConstValueShape).

Subtype of AbstractValueShape must support eltype, size and totalndof.

source

ShapesOfVariables.ArrayShape — Type.

ArrayShape{T,N} <: AbstractValueShape

Describes the shape of N-dimensional arrays of type T and a given size.

Constructor:

ArrayShape{T}(dims::NTuple{N,Integer}) where {T,N}
ArrayShape{T}(dims::Integer...) where {T}

e.g.

shape = ArrayShape{Real}(2, 3)

Array shapes can be used to instantiate array of the given shape, e.g.

size(Array(undef, shape)) == (2, 3)
size(ElasticArrays.ElasticArray(undef, shape)) == (2, 3)

If the element type of the shape of an abstract type of union, ShapesOfVariables.default_datatype will be used to determine a suitable more specific type, if possible:

eltype(Array(undef, shape)) == Float64

source

ShapesOfVariables.ConstValueShape — Type.

ConstValueShape{T} <: AbstractValueShape

A ConstValueShape describes the shape of constant values of type T.

Constructor:

ConstValueShape(value)

value may be of arbitrary type, e.g. a constant scalar value or array:

ConstValueShape(4.2),
ConstValueShape([11 21; 12 22]),

Shapes of constant values have zero degrees of freedom ((see totalndof).

source

ShapesOfVariables.ScalarShape — Type.

ScalarShape{T} <: AbstractValueShape

An ScalarShape describes the shape of scalar values of a given type.

Constructor:

ScalarShape{T::Type}()

T may be an abstract type of Union, or a specific type, e.g.

ScalarShape{Integer}()
ScalarShape{Real}()
ScalarShape{Complex}()
ScalarShape{Float32}()

Scalar shapes may have a total number of degrees of freedom (see totalndof) greater than one, e.g. shapes of complex-valued scalars:

totalndof(ScalarShape{Real}()) == 1
totalndof(ScalarShape{Complex}()) == 2

source

ShapesOfVariables.VarShapes — Type.

VarShapes{N,AC}

Defines the shapes of an ordered set of variables (resp. parameters, arguments, etc.). This forms the basis of viewing the content of all variables in a dual way as a NamedTuple and as a flattened vectors.

Scalar values have shape (), array values have shape (dim1, dim2, ...).

Constructors:

VarShapes(name1 = shape1, ...)
VarShapes(varshapes::NamedTuple)

e.g.

varshapes = VarShapes(
    a = ArrayShape{Real}(2, 3),
    b = ScalarShape{Real}(),
    c = ArrayShape{Real}(4)
)

Use

(varshapes::VarShapes)(data::AbstractVector)::NamedTuple

to get correctly named and shaped views into a vector containing the flattened values of all variables. In return,

Base.Vector{T}(::UndefInitializer, varshapes::VarShapes)
Base.Vector(::UndefInitializer, varshapes::VarShapes)

will create a suitable uninitialized vector to hold such flattened data for a given set of variables. If no type T is given, a suitable non-abstract type will be chosen automatically via nonabstract_eltype(varshapes).

When dealing with multiple vectors of flattened data,

(varshapes::VarShapes)(
    data::ArrayOfArrays.AbstractVectorOfSimilarVectors
)::NamedTuple

creates a view of a vector of flattened data vectors as a table with the variable names as column names and the (possibly array-shaped) variable value views as entries. In return,

ArraysOfArrays.VectorOfSimilarVectors{T}(varshapes::VarShapes)
ArraysOfArrays.VectorOfSimilarVectors(varshapes::VarShapes)

will create a suitable vector (of length zero) of vectors to hold flattened value data. The result will be a VectorOfSimilarVectors wrapped around a 2-dimensional ElasticArray. Internally all data is stored in a single flat Vector{T}.

Example:

varshapes = VarShapes(
    a = ScalarShape{Real}(),
    b = ArrayShape{Real}(2, 3),
    c = ConstValueShape(42)
)
data = VectorOfSimilarVectors{Float64}(varshapes)
resize!(data, 10)
rand!(flatview(data))
table = TypedTables.Table(varshapes(data))
fill!(table.a, 4.2)
all(x -> x == 4.2, view(flatview(data), 1, :))

source

ShapesOfVariables.totalndof — Function.

totalndof(shape::AbstractValueShape)

Get the total number of degrees of freedom of values having the given shape.

Equivalent to the size of the array required when flattening values of this shape into an array of real numbers, without including any constant values.

source

ShapesOfVariables.default_datatype — Function.

ShapesOfVariables.default_datatype(T::Type)

Return a default specific type U that is more specific than T, with U <: T.

e.g.

ShapesOfVariables.default_datatype(Real) == Float64
ShapesOfVariables.default_datatype(Complex) == Complex{Float64}

source